pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vidforge"
version = "0.1.0"
description = "Long-form text-to-video and image-to-video rendering toolkit"
authors = [{ name = "VidForge Contributors" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12",
    "pydantic>=1.10",
    "numpy>=1.24",
    "rich>=13.0",
    "opencv-python>=4.8",
    "moviepy>=1.0",
    "soundfile>=0.12",
    "tqdm>=4.66",
    "einops>=0.7",
    "ffmpeg-python>=0.2",
]

[project.optional-dependencies]
models = [
    "diffusers>=0.27",
    "xformers>=0.0.26",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.6",
]

[project.entry-points."console_scripts"]
vidforge = "vidforge.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["vidforge*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
